Set --single-directory=true on informer gen
