+++ hack/update-client-gen.sh
@@ -11,20 +11,21 @@ BINDIR=${REPO_ROOT}/bin
 
 # Generate the versioned clientset (pkg/client/clientset_generated/clientset)
 ${BINDIR}/client-gen "$@" \
-		  --input-base "github.com/jetstack-experimental/cert-manager/pkg/apis/" \
-		  --input "certmanager/v1alpha1" \
-	      --clientset-path "github.com/jetstack-experimental/cert-manager/pkg/" \
-	      --clientset-name "client" \
-	      --go-header-file "${GOPATH}/src/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt"
+          --input-base "github.com/jetstack-experimental/cert-manager/pkg/apis/" \
+          --input "certmanager/v1alpha1" \
+          --clientset-path "github.com/jetstack-experimental/cert-manager/pkg/" \
+          --clientset-name "client" \
+          --go-header-file "${GOPATH}/src/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt"
 # generate lister
 ${BINDIR}/lister-gen "$@" \
-	      --input-dirs="github.com/jetstack-experimental/cert-manager/pkg/apis/certmanager/v1alpha1" \
-	      --output-package "github.com/jetstack-experimental/cert-manager/pkg/listers" \
-	      --go-header-file "${GOPATH}/src/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt"
+          --input-dirs="github.com/jetstack-experimental/cert-manager/pkg/apis/certmanager/v1alpha1" \
+          --output-package "github.com/jetstack-experimental/cert-manager/pkg/listers" \
+          --go-header-file "${GOPATH}/src/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt"
 # generate informer
 ${BINDIR}/informer-gen "$@" \
-	      --go-header-file "${GOPATH}/src/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt" \
-	      --input-dirs "github.com/jetstack-experimental/cert-manager/pkg/apis/certmanager/v1alpha1" \
-	      --versioned-clientset-package "github.com/jetstack-experimental/cert-manager/pkg/client" \
-	      --listers-package "github.com/jetstack-experimental/cert-manager/pkg/listers" \
-	      --output-package "github.com/jetstack-experimental/cert-manager/pkg/informers"
+          --go-header-file "${GOPATH}/src/github.com/kubernetes/repo-infra/verify/boilerplate/boilerplate.go.txt" \
+          --input-dirs "github.com/jetstack-experimental/cert-manager/pkg/apis/certmanager/v1alpha1" \
+          --versioned-clientset-package "github.com/jetstack-experimental/cert-manager/pkg/client" \
+          --listers-package "github.com/jetstack-experimental/cert-manager/pkg/listers" \
+          --output-package "github.com/jetstack-experimental/cert-manager/pkg/informers" \
+          --single-directory
