+++ pkg/informers/certmanager/interface.go
@@ -19,8 +19,8 @@ limitations under the License.
 package certmanager
 
 import (
-	v1alpha1 "github.com/jetstack-experimental/cert-manager/pkg/informers/externalversions/certmanager/v1alpha1"
-	internalinterfaces "github.com/jetstack-experimental/cert-manager/pkg/informers/externalversions/internalinterfaces"
+	v1alpha1 "github.com/jetstack-experimental/cert-manager/pkg/informers/certmanager/v1alpha1"
+	internalinterfaces "github.com/jetstack-experimental/cert-manager/pkg/informers/internalinterfaces"
 )
 
 // Interface provides access to each of this group's versions.
