+++ pkg/informers/certmanager/v1alpha1/interface.go
@@ -19,7 +19,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	internalinterfaces "github.com/jetstack-experimental/cert-manager/pkg/informers/externalversions/internalinterfaces"
+	internalinterfaces "github.com/jetstack-experimental/cert-manager/pkg/informers/internalinterfaces"
 )
 
 // Interface provides access to all the informers in this group version.
