+++ pkg/informers/factory.go
@@ -16,12 +16,12 @@ limitations under the License.
 
 // This file was automatically generated by informer-gen
 
-package externalversions
+package informers
 
 import (
 	client "github.com/jetstack-experimental/cert-manager/pkg/client"
-	certmanager "github.com/jetstack-experimental/cert-manager/pkg/informers/externalversions/certmanager"
-	internalinterfaces "github.com/jetstack-experimental/cert-manager/pkg/informers/externalversions/internalinterfaces"
+	certmanager "github.com/jetstack-experimental/cert-manager/pkg/informers/certmanager"
+	internalinterfaces "github.com/jetstack-experimental/cert-manager/pkg/informers/internalinterfaces"
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	cache "k8s.io/client-go/tools/cache"
