+++ pkg/informers/generic.go
@@ -16,7 +16,7 @@ limitations under the License.
 
 // This file was automatically generated by informer-gen
 
-package externalversions
+package informers
 
 import (
 	"fmt"
